Add support for backwards migrations to migrations.Distribute
Since Citus 9.5 you can now undistribute a table. This allows
us to easily add support for backwards migrations.

This also serves as a regression test for the problem that [PR #95][1]
fixes. Inbetween test runs we now clean up the whole database and the
migrations are run for each Django/Python version combination that we
run.

1: https://github.com/citusdata/django-multitenant/pull/95