Update versions throughout the repository (#104)
This tests that we support newer Django, Python and Citus versions and
drops support for Django versions that have reached upstream EOL.

The only EOL version that we still test against is django 1.11, because
that's the last one to support Python 2.7. I'm pretty sure some people
still use this even though they shouldn't. So let's try to keep
supporting it for now.