Add support for backwards migrations to migrations.Distribute (#106)
Since Citus 9.5 you can now undistribute a distributed table. This 
allows us to easily add support for backwards migrations.

To test this we now run all of the migrations backwards (except for 
one). Most of the raw sql, I did not add a reverse migrations for. 
This seems unnecessary, because a migrations down the line will delete 
the whole table anyway.

Because we run these backwards migrations, this also serves as a 
regression test for the problem that [PR #95][1] fixes. Inbetween 
test runs we now clean up the whole database and the migrations 
are run for each Django/Python version combination that we run.
So if newer Django versions break the migrations we'll notice now.

[1]: https://github.com/citusdata/django-multitenant/pull/95