Drop support for EOL Django 3.1
When Django 4.0 was released the Django team dropped support for Django 3.1.
In the next release we'll only support versions that are supported
by upstream. So we're dropping support for Django 3.1. Sadly we cannot
remove any version checks since they all impact Django 2.2 as well,
which is still supported because it's an LTS release.