Add random PK for fixtures
The static primary keys are being overwritten by a recent migration that
seeded data. Removing the primary keys from these fixtures results in
failures of the UtilsTest test cases; this appears to be due to Django
generating UUIDs for the ID field after a model with a model with a UUID
ID field - such as the Organization, OrganizationLocation or Record -
is created. This is a stopgap solution.